refactor: regenerate permission proxies
