+++ npm/ng-packs/packages/permission-management/src/lib/proxy/index.ts
