+++ npm/ng-packs/packages/permission-management/src/lib/proxy/models.ts
