+++ npm/ng-packs/packages/permission-management/src/public-api.ts
@@ -9,4 +9,4 @@ export * from './lib/enums/components';
 export * from './lib/models';
 export * from './lib/services';
 export * from './lib/states';
-export * from './lib/proxy/permission-management';
+export * from './lib/proxy';
